refactor: put `EditableTable.ts` in `src/Editor/`
- Updated `Table.ts` to utilize `ImageHandler`, `ExportHandler`, and `ColumnHelper` for improved modularity and code organization.
- Extracted image handling and rendering logic into a new file `ImageHandler.ts`, to manage image-related functionalities independently.
- Moved exporting logic to a separate file `ExportHandler.ts`, for managing CSV and Excel export functionalities.
- Created `ColumnHelper.ts` to encapsulate column-related operations, providing methods for generating columns, column widths, row indices, and MEI data.

Refs: #73
